Delete umabigous src folder, move stuff to utils
